Custom User Model for Urbvan ----------------------------
This implements a Custom User Model for the Urbvan app for easily
extending the User Model and avoid the overhead of using
Django's Auth Groups by defining an User Type in it (and updating
the admin interface).

The implementation uses the same table as Django Auth for easing the
transition to it, however, if you already have ran the migrations you
*have* to:
- Truncate the Django's migrations table, i.e. in sqlite3:
	> python manage.py dbshell
	>> DELETE FROM django_migrations;
- Fake all migrations:
	> python manage.py migrate --fake
- Set the migration state to the initial in Users:
	> python manage.py migrate users 0001 --fake
- Apply the pending migration in Users
	> python manage.py migrate users

If not, everything should go smoothly.

Note that from now on, any reference to the User model should be done
either with django.contrib.auth.get_user_model() or
settings.AUTH_USER_MODEL.