Add a helper function 'create_id_for_MODEL' to each model that uses create_id() for generating its id to fix the creation of a migration each time makemigrations is called (using parameters in callables makes the function to be evaluated instead of being passed as a function, leading to the trigger of the field altering by the migrations generator).
Warning: A migration might be generated for each model the next
time makemigrations is called with a new default id. It can be
safely removed or ignored and it won't be generated anymore.
This can be triggered to avoid surprises in a future (call
makemigrations and choose either to delete or ignore the generated
ones - provided that no changes to the models were done)