CRUD Model Permissions ----------------------
CRUD Model Permissions is a Group / Role - based permission system for
CRUD actions on Models which requires no additional queries to the
database (no database overhead, no penality on performance).

It is intended to be used in the Models' CRUD endpoints of the REST API.

Read the full documentation in apps.urbvan_framework.permissions.

The system is 'end-to-end' tested with the Location CRUD endpoints.